+++ CHANGELOG.md
@@ -26,6 +26,9 @@ All significant changes to this project will be documented in this file.
 #### Fixed
 - Fixed `logTrace` when no trace level is passed.  It's now the correct default value of 1 instead of 4 (issue #58).
 
+#### Removed 
+- Removed  `TraceLogTestHarness`  module.
+
 ## [4.0.1](https://github.com/tonystone/tracelog/tree/4.0.1)
 
 #### Fixed